HSSFN/requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
einops==0.8.1
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
graphviz==0.20.3
h5py==3.13.0
hdf5storage==0.1.19
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.2.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
propcache==0.3.0
psutil==7.0.0
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
spectral==0.23.1
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.21.1
tomli==2.2.1
torch ==2.5.1
torch-geometric==2.6.1
torch-summary==1.4.5
torchviz==0.0.3
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
